Golf

+++ src/core/Ekubo.sol
@@ -59,6 +59,10 @@ interface IEkuboCallbacks {
 }
 
 library UnsafeEkuboCore {
+    /// The `amountSpecified` as well as both `delta`'s are `int256` for contract size savings. If
+    /// `amountSpecified` is not a clean, signed, 128-bit value, the call will revert inside the ABI
+    /// decoding in `CORE`. The `delta`'s are guaranteed clean by the returndata encoding of `CORE`,
+    /// but we keep them as `int256` so as not to duplicate any work.
     function unsafeSwap(
         IEkuboCore core,
         PoolKey memory poolKey,
@@ -66,7 +70,7 @@ library UnsafeEkuboCore {
         bool isToken1,
         SqrtRatio sqrtRatioLimit,
         uint256 skipAhead
-    ) internal returns (int128 delta0, int128 delta1) {
+    ) internal returns (int256 delta0, int256 delta1) {
         assembly ("memory-safe") {
             let ptr := mload(0x40)
 
@@ -378,7 +382,7 @@ abstract contract Ekubo is SettlerAbstract {
                     uint96(isToken1.ternary(uint256(79227682466138141934206691491), uint256(4611797791050542631)))
                 );
                 (int256 delta0, int256 delta1) = IEkuboCore(msg.sender).unsafeSwap(
-                    poolKey, int128(amountSpecified), isToken1, sqrtRatio, skipAhead
+                    poolKey, amountSpecified, isToken1, sqrtRatio, skipAhead
                 );
                 // Ekubo's sign convention here is backwards compared to UniV4/BalV3/PancakeInfinity
                 // `settledSellAmount` is positive, `settledBuyAmount` is negative. So the use of
